SGLCN.README.md
SeagullCoin is a Seagull currency that acts as gold in the Bored Seagull Club Metaverse with a supply capped at 589 million tokens.

This is considered the layer 1 solution of our token economy.

With NFT and a SGLCN-X20 standard. This allows participants in our metaverse to quickly exchange with a gold value to other Seagull Assets and main source of liquidity for our participants.